# ddtrace/tracer: optimize applyPPROFLabels for go1.24