--- a/ddtrace/tracer/tracer.go
+++ b/ddtrace/tracer/tracer.go
@@ -681,25 +681,25 @@ func (t *tracer) StartSpan(operationName string, options ...ddtrace.StartSpanOpt
 // found in ctx are restored. Additionally, this func informs the profiler how
 // many times each endpoint is called.
 func (t *tracer) applyPPROFLabels(ctx gocontext.Context, span *span) {
-	var labels []string
+	// Important: The label keys are ordered alphabetically to take advantage of
+	// an upstream optimization that landed in go1.24.  This results in ~10%
+	// better performance on BenchmarkStartSpan. See
+	// https://go-review.googlesource.com/c/go/+/574516 for more information.
+	labels := make([]string, 0, 3*2 /* 3 key value pairs */)
+	localRootSpan := span.root()
+	if t.config.profilerHotspots && localRootSpan != nil {
+		labels = append(labels, traceprof.LocalRootSpanID, strconv.FormatUint(localRootSpan.SpanID, 10))
+	}
 	if t.config.profilerHotspots {
-		// allocate the max-length slice to avoid growing it later
-		labels = make([]string, 0, 6)
 		labels = append(labels, traceprof.SpanID, strconv.FormatUint(span.SpanID, 10))
 	}
-	// nil checks might not be needed, but better be safe than sorry
-	if localRootSpan := span.root(); localRootSpan != nil {
-		if t.config.profilerHotspots {
-			labels = append(labels, traceprof.LocalRootSpanID, strconv.FormatUint(localRootSpan.SpanID, 10))
-		}
-		if t.config.profilerEndpoints && spanResourcePIISafe(localRootSpan) {
-			labels = append(labels, traceprof.TraceEndpoint, localRootSpan.Resource)
-			if span == localRootSpan {
-				// Inform the profiler of endpoint hits. This is used for the unit of
-				// work feature. We can't use APM stats for this since the stats don't
-				// have enough cardinality (e.g. runtime-id tags are missing).
-				traceprof.GlobalEndpointCounter().Inc(localRootSpan.Resource)
-			}
+	if t.config.profilerEndpoints && spanResourcePIISafe(localRootSpan) && localRootSpan != nil {
+		labels = append(labels, traceprof.TraceEndpoint, localRootSpan.Resource)
+		if span == localRootSpan {
+			// Inform the profiler of endpoint hits. This is used for the unit of
+			// work feature. We can't use APM stats for this since the stats don't
+			// have enough cardinality (e.g. runtime-id tags are missing).
+			traceprof.GlobalEndpointCounter().Inc(localRootSpan.Resource)
 		}
 	}
 	if len(labels) > 0 {