@felixge 2025-03-02
> Found 0 performance improvements and 0 performance regressions! Performance is the same for 60 metrics, 2 unstable metrics.

This is a red herring. The benchmark platform uses go1.23, but this patch relies on an upstream patch in go1.24.

To double check, I also ran the benchmarks on a amd64 workspace machine, see results below:

```
goos: linux
goarch: amd64
pkg: gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer
cpu: Intel(R) Xeon(R) Platinum 8175M CPU @ 2.50GHz
             │  base.txt   │             patch1.txt              │
             │   sec/op    │   sec/op     vs base                │
StartSpan-16   1.786µ ± 5%   1.602µ ± 2%  -10.28% (p=0.000 n=10)
```